fix: Sticky navbar, move modals to own components, tidy up CSS, use store for `passed_login_with_creds`
